+++ src/Beef.Core/Events/EventData.cs
@@ -19,7 +19,17 @@ namespace Beef.Events
         /// <param name="action">The event action.</param>
         /// <returns>The <see cref="EventData"/>.</returns>
         public static EventData CreateEvent(string subject, string? action = null) 
-            => new EventData { Subject = Check.NotEmpty(subject, nameof(subject)), Action = action };
+            => new() { Subject = Check.NotEmpty(subject, nameof(subject)), Action = action };
+
+        /// <summary>
+        /// Creates an <see cref="EventData"/> instance with no <see cref="Key"/>.
+        /// </summary>
+        /// <param name="source">The event source.</param>
+        /// <param name="subject">The event subject.</param>
+        /// <param name="action">The event action.</param>
+        /// <returns>The <see cref="EventData"/>.</returns>
+        public static EventData CreateEvent(Uri source, string subject, string? action = null)
+            => new() { Source = Check.NotNull(source, nameof(source)), Subject = Check.NotEmpty(subject, nameof(subject)), Action = action };
 
         /// <summary>
         /// Creates an <see cref="EventData"/> instance with the specified <see cref="Key"/>.
@@ -29,7 +39,18 @@ namespace Beef.Events
         /// <param name="key">The event key.</param>
         /// <returns>The <see cref="EventData"/>.</returns>
         public static EventData CreateEvent(string subject, string? action = null, params IComparable?[] key) 
-            => new EventData { Subject = Check.NotEmpty(subject, nameof(subject)), Action = action, Key = key.Length == 1 ? (object?)key[0] : key };
+            => new() { Subject = Check.NotEmpty(subject, nameof(subject)), Action = action, Key = key.Length == 1 ? (object?)key[0] : key };
+
+        /// <summary>
+        /// Creates an <see cref="EventData"/> instance with the specified <see cref="Key"/>.
+        /// </summary>
+        /// <param name="source">The event source.</param>
+        /// <param name="subject">The event subject.</param>
+        /// <param name="action">The event action.</param>
+        /// <param name="key">The event key.</param>
+        /// <returns>The <see cref="EventData"/>.</returns>
+        public static EventData CreateEvent(Uri source, string subject, string? action = null, params IComparable?[] key)
+            => new() { Source = Check.NotNull(source, nameof(source)), Subject = Check.NotEmpty(subject, nameof(subject)), Action = action, Key = key.Length == 1 ? (object?)key[0] : key };
 
         /// <summary>
         /// Creates an <see cref="EventData"/> instance using the <paramref name="value"/> (infers the <see cref="Key"/> from either <see cref="IIdentifier"/> or <see cref="IUniqueKey"/>).
@@ -68,6 +89,22 @@ namespace Beef.Events
         }
 
         /// <summary>
+        /// Creates an <see cref="EventData"/> instance using the <paramref name="value"/> (infers the <see cref="Key"/> from either <see cref="IIdentifier"/> or <see cref="IUniqueKey"/>).
+        /// </summary>
+        /// <typeparam name="T">The value <see cref="Type"/>.</typeparam>
+        /// <param name="value">The event value</param>
+        /// <param name="source">The event source.</param>
+        /// <param name="subject">The event subject.</param>
+        /// <param name="action">The event action.</param>
+        /// <returns>The <see cref="EventData"/>.</returns>
+        public static EventData<T> CreateValueEvent<T>(T value, Uri source, string subject, string? action = null) where T : class
+        {
+            var ed = CreateValueEvent(value, subject, action);
+            ed.Source = Check.NotNull(source, nameof(source));
+            return ed;
+        }
+
+        /// <summary>
         /// Creates an <see cref="EventData"/> instance with the specified <paramref name="value"/> <see cref="Key"/>.
         /// </summary>
         /// <typeparam name="T">The value <see cref="Type"/>.</typeparam>
@@ -77,7 +114,24 @@ namespace Beef.Events
         /// <param name="key">The event key.</param>
         /// <returns>The <see cref="EventData"/>.</returns>
         public static EventData<T> CreateValueEvent<T>(T value, string subject, string? action = null, params IComparable?[] key) 
-            => new EventData<T> { Value = value, Subject = Check.NotEmpty(subject, nameof(subject)), Action = action, Key = key.Length == 1 ? (object?)key[0] : key };
+            => new() { Value = value, Subject = Check.NotEmpty(subject, nameof(subject)), Action = action, Key = key.Length == 1 ? (object?)key[0] : key };
+
+        /// <summary>
+        /// Creates an <see cref="EventData"/> instance with the specified <paramref name="value"/> <see cref="Key"/>.
+        /// </summary>
+        /// <typeparam name="T">The value <see cref="Type"/>.</typeparam>
+        /// <param name="value">The event value</param>
+        /// <param name="source">The event source.</param>
+        /// <param name="subject">The event subject.</param>
+        /// <param name="action">The event action.</param>
+        /// <param name="key">The event key.</param>
+        /// <returns>The <see cref="EventData"/>.</returns>
+        public static EventData<T> CreateValueEvent<T>(T value, Uri source, string subject, string? action = null, params IComparable?[] key)
+        {
+            var ed = CreateValueEvent(value, subject, action, key);
+            ed.Source = Check.NotNull(source, nameof(source));
+            return ed;
+        }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="EventData"/> class defaulting as applicable using the equivalent <see cref="ExecutionContext"/> <see cref="ExecutionContext.Current"/> values.
@@ -110,6 +164,12 @@ namespace Beef.Events
         /// </summary>
         [JsonProperty("tenantId", DefaultValueHandling = DefaultValueHandling.Ignore)]
         public Guid? TenantId { get; set; }
+
+        /// <summary>
+        /// Gets or sets the event source <see cref="Uri"/> (describes the event producer).
+        /// </summary>
+        [JsonProperty("source", DefaultValueHandling = DefaultValueHandling.Ignore)]
+        public Uri? Source { get; set; }
 
         /// <summary>
         /// Gets or sets the event subject (the name should use the '.' character to denote paths).