+++ src/Beef.Core/Events/EventStringFormat.cs
@@ -1,0 +1,25 @@
+﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/Beef
+
+namespace Beef.Events
+{
+    /// <summary>
+    /// Defines the event string format.
+    /// </summary>
+    public enum EventStringFormat
+    {
+        /// <summary>
+        /// No formatting; as-is (default).
+        /// </summary>
+        None,
+
+        /// <summary>
+        /// Format as upper case.
+        /// </summary>
+        Uppercase,
+
+        /// <summary>
+        /// Format as lower case.
+        /// </summary>
+        Lowercase
+    }
+}