+++ tools/Beef.CodeGen.Core/CodeGenFileManager.cs
@@ -21,22 +21,22 @@ namespace Beef.CodeGen
         /// <summary>
         /// Gets the list of supported <see cref="CommandType.Entity"/> filenames (will search in order specified).
         /// </summary>
-        public static List<string> EntityFilenames { get; } = new List<string>(new string[] { "entity.beef.yaml", "entity.beef.json", "entity.beef.xml", "{{Company}}.{{AppName}}.xml" });
+        public static List<string> EntityFilenames { get; } = new List<string>(new string[] { "entity.beef.yaml", "entity.beef.yml", "entity.beef.json", "entity.beef.xml", "{{Company}}.{{AppName}}.xml" });
 
         /// <summary>
         /// Gets the list of supported <see cref="CommandType.RefData"/> filenames (will search in order specified).
         /// </summary>
-        public static List<string> RefDataFilenames { get; } = new List<string>(new string[] { "refdata.beef.yaml", "refdata.beef.json", "refdata.beef.xml", "{{Company}}.RefData.xml" });
+        public static List<string> RefDataFilenames { get; } = new List<string>(new string[] { "refdata.beef.yaml", "refdata.beef.yml", "refdata.beef.json", "refdata.beef.xml", "{{Company}}.RefData.xml" });
 
         /// <summary>
         /// Gets the list of supported <see cref="CommandType.RefData"/> filenames (will search in order specified).
         /// </summary>
-        public static List<string> DataModelFilenames { get; } = new List<string>(new string[] { "datamodel.beef.yaml", "datamodel.beef.json", "datamodel.beef.xml", "{{Company}}.{{AppName}}.DataModel.xml" });
+        public static List<string> DataModelFilenames { get; } = new List<string>(new string[] { "datamodel.beef.yaml", "datamodel.beef.yml", "datamodel.beef.json", "datamodel.beef.xml", "{{Company}}.{{AppName}}.DataModel.xml" });
 
         /// <summary>
         /// Gets the list of supported <see cref="CommandType.Database"/> filenames (will search in order specified).
         /// </summary>
-        public static List<string> DatabaseFilenames { get; } = new List<string>(new string[] { "database.beef.yaml", "database.beef.json", "database.beef.xml", "{{Company}}.{{AppName}}.Database.xml" });
+        public static List<string> DatabaseFilenames { get; } = new List<string>(new string[] { "database.beef.yaml", "database.beef.yml", "database.beef.json", "database.beef.xml", "{{Company}}.{{AppName}}.Database.xml" });
 
         /// <summary>
         /// Get the configuration filename.
@@ -136,9 +136,7 @@ namespace Beef.CodeGen
 
                 logger.LogInformation(string.Empty);
             }
-#pragma warning disable CA1031 // Do not catch general exception types; is OK.
             catch (Exception ex)
-#pragma warning restore CA1031 
             {
                 logger.LogError(ex.Message);
                 logger.LogInformation(string.Empty);