+++ tools/Beef.CodeGen.Core/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Represents the **NuGet** versions.
 
+## v4.1.23
+- *Enhancement:* Added support for YAML files with a `.yml` suffix.
+- *Enhancement:* Added support for `EventDasta.Source` to both entity and database (CDC) code-gen.
+
 ## v4.1.22
 - *Enhancement:* Added standardized identifier mapping for change-data-capture (CDC) from local (internal) to global (external) where required.
 - *Enhancement:* Added additional statistics information to console output.